+++ src/test/java/UserDaoEqualityTest.java
@@ -2,7 +2,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.AnnotationConfigApplicationContext;
 import org.springframework.test.annotation.DirtiesContext;
-import user.domain.UserDao;
+import user.dao.UserDao;
 
 import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
